+++ app/Console/Framework/AES/NewAESCommand.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Framework\AES;
 
+use App\Traits\Framework\AES;
 use App\Traits\Framework\ConsoleOutput;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -9,7 +10,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class NewAESCommand extends Command {
 
-    use ConsoleOutput;
+    use ConsoleOutput, AES;
 
 	protected static $defaultName = "aes:new";
 
@@ -27,22 +28,8 @@ class NewAESCommand extends Command {
     }
 
     protected function execute(InputInterface $input, OutputInterface $output) {
-        $generateKeys = function() {
-            $items = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789@-_/*{}[].,#$&';
-            $bytes = random_bytes(16);
-            $longitud = strlen($items);
-            $key = '';
-
-            for ($i = 0; $i < 16; $i++) {
-                $indice = ord($bytes[$i]) % $longitud;
-                $key .= $items[$indice];
-            }
-
-            return $key;
-        };
-
-        $output->writeln($this->warningOutput("\t>>  AES KEY: {$generateKeys()}"));
-        $output->writeln($this->warningOutput("\t>>  AES IV: {$generateKeys()}"));
+        $output->writeln($this->warningOutput("\t>>  AES KEY: {$this->generateKeys()}"));
+        $output->writeln($this->warningOutput("\t>>  AES IV: {$this->generateKeys()}"));
         $output->writeln($this->successOutput("\t>>  Keys created successfully"));
         return Command::SUCCESS;
     }