+++ app/Traits/Framework/AES.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Traits\Framework;
+
+trait AES {
+
+    public function generateKeys(): string {
+        $items = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789@-_/*{}[].,#$&';
+        $bytes = random_bytes(16);
+        $longitud = strlen($items);
+        $key = '';
+
+        for ($i = 0; $i < 16; $i++) {
+            $indice = ord($bytes[$i]) % $longitud;
+            $key .= $items[$indice];
+        }
+
+        return $key;
+    }
+
+}