+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [v14.29.0](https://github.com/Lion-Packages/framework/compare/v14.28.1...v14.29.0) (2023-07-20)
 
+### Added
+- added trait to generate keys for AES
+
 ### Changed
 - output format for rules have been modified
 - RSA route settings have been updated